物流狀態，新增表單的 啟用狀態按鈕 Ok，且確定可以新增，新增完畢後，會跳轉回物六列表

+++ Team1.VitalBridge.BackStage/Models/ViewModels/shipViewModel.cs
@@ -15,7 +15,7 @@ namespace Team1.VitalBridge.BackStage.Models.ViewModels
         [Required]
 		[DataType(DataType.Currency)]
 		[Display(Name = "運費")]
-        public decimal ShipCost { get; set; }
+        public int ShipCost { get; set; }
 
         [Display(Name = "啟用狀態")]
         public bool? IsActive { get; set; }
